+++ src/utils/loader.js
@@ -6,21 +6,21 @@ export class LoaderClass {
     data: false,
     page: false,
     reload: false,
-    images: false
+    images: false,
   };
   delay = {
     firstload: 0,
     default: 250,
-    reload: 250
+    reload: 250,
   };
   timeout = {
-    time: null
+    time: null,
   };
   images = {
     check: true,
     init: false,
     list: null,
-    loaded: 0
+    loaded: 0,
   };
   constructor(obj) {
     this.loadcheck.time = false;
@@ -115,35 +115,35 @@ export class LoaderClass {
   imageload() {
     if (this.images.check) {
       if (this.images.list && this.images.list.length > 0) {
-        this.images.list.forEach(image => {
+        this.images.list.forEach((image) => {
           //ADD CONDITIONAL TO CHECK IF IMAGE HAS VALID SRC
           if (
             image.src !== null &&
             image.src !== '' &&
             image.src !== undefined
           ) {
-            const eachimageloaded = e => {
+            const eachimageloaded = (e) => {
               e.target.removeEventListener('load', eachimageloaded, false);
               e.target.removeEventListener('error', eachimageerror, false);
               this.images.loaded++;
               this.imageloadfinish();
             };
-            const eachimageerror = e => {
+            const eachimageerror = (e) => {
               e.target.removeEventListener('load', eachimageloaded, false);
               e.target.removeEventListener('error', eachimageerror, false);
               this.images.loaded++;
-              console.log('fail to load',e.target);
+              console.log('fail to load', e.target);
               this.imageloadfinish();
             };
-            if (image.complete) {
+            // Check if image is already loaded or set to lazy.
+            if (image.complete || image.loading === 'lazy' || image.ariaHidden) {
               this.images.loaded++;
               this.imageloadfinish();
             } else {
               image.addEventListener('load', eachimageloaded, false);
               image.addEventListener('error', eachimageerror, false);
             }
-          }
-          else{
+          } else {
             this.images.loaded++;
             this.imageloadfinish();
           }